requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6

# Database
SQLAlchemy==2.0.36
alembic==1.13.2
pymysql==1.1.1

# Auth / security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Settings / utils
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0

